requirements.txt
pandas
pyarrow
sqlalchemy
numpy
sphinx
numpy
sphinxcontrib-bibtex
sphinx-gallery
sphinx-prompt
sphinx-copybutton
pydata-sphinx-theme
scipy
ruff
isort
blackdoc
python-dotenv
sshtunnel
matplotlib
psycopg2
polars
connectorx
alive_progress